Move tri-mesh collision module to geometry layer
Relocate TriMeshCollisionInfo, TriMeshCollisionDetector, and the
filter-list helpers from newton/_src/solvers/vbd/ to
newton/_src/geometry/ so the sim layer (CollisionPipeline, Contacts)
can reference them without importing from the solvers layer. Keep the
old module as a re-export; switch VBD imports to the new path.

No behavior change: test_collision_cloth (71) and vbd (210) suites
pass unchanged on GPU.
